% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bin_vector}
\alias{bin_vector}
\title{Bin vectors}
\usage{
bin_vector(x, width = 1, origin = 0, weight = NULL, closed = c("right",
  "left"), pad = FALSE, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector (or S3 class based on a numeric) to bin.}

\item{width}{The width of a bin. Must be positive.}

\item{origin}{The location of the left-most bin edge. Any values smaller
than the \code{origin} will be treated as if they are missing.}

\item{weight}{If specified, an integer vector of the same length as \code{x}
giving weights. If weights are provided, the weights in each bin are
summed, rather than just counting the number of observations.}

\item{closed}{One of \code{"right"} or \code{"left"} indicating whether the
bin interval is left-closed (i.e. [a, b)), or right-closed (i.e. (a, b]).}

\item{pad}{A logical indicating whether the bins should be padded to include
an empty bin on each side. This is useful for frequency polygons which
need to go back down to zero at either end of the range.}
}
\description{
A generic and several implementations for binning vectors.
}
\examples{
x <- runif(1e6)
bin_vector(x, 0.1)

# Performance scales linearly with the size of x, and the number
# of bins has limited impact
x <- runif(1e7)
system.time(bin_vector(x))
system.time(bin_vector(x, width = 1 / 100))
system.time(bin_vector(x, width = 1 / 1e5))
}
\keyword{internal}

