% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geometry_flip}
\alias{geometry_flip}
\title{Flip positions around an axis.}
\usage{
geometry_flip(geom, dir = c("x", "y"))
}
\arguments{
\item{geom}{A geometry data frame.}

\item{dir}{Direction in which to flip. One of "x" and "y".}
}
\description{
Flip positions around an axis.
}
\examples{
scatter_ex \%>\% plot()
scatter_ex \%>\% geometry_flip() \%>\% plot()
scatter_ex \%>\% geometry_flip("y") \%>\% plot()

histogram_ex \%>\% plot()
histogram_ex \%>\% geometry_flip() \%>\% plot()
histogram_ex \%>\% geometry_flip("y") \%>\% plot()
histogram_ex \%>\% geometry_flip("y") \%>\% geometry_rotate() \%>\% plot()
}

