% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geometry_pointificate}
\alias{geometry_pointificate}
\title{Convert complex geometries in to points, paths and polygons.}
\usage{
geometry_pointificate(geom, ...)
}
\arguments{
\item{geom}{A geometry data frame.}

\item{...}{Additional arguments passed on to methods.
\itemize{
 \item{\code{geometry_polygon}: use \code{close = TRUE} to "close" the
   polygon by putting the first point at the end.}
}}
}
\description{
Convert complex geometries in to points, paths and polygons.
}
\examples{
x <- seq(0, 4 * pi, length = 100)
df <- data.frame(x = x, y = sin(x))
rib <- render_ribbon(df, ~x, ~y - 1, ~ y + 1)
rib \%>\% plot()
rib \%>\% geometry_pointificate() \%>\% plot()
rib \%>\% geometry_pointificate() \%>\% geometry_flip() \%>\% plot()

df <- expand.grid(x = 1:3, y = 1:3)
df$z <- runif(9, pi, 2 * pi)
arc <- df \%>\% render_arc(~x, ~y, 0, 0.35, 0, ~z)
arc \%>\% plot()
arc \%>\% geometry_pointificate() \%>\% plot()

histogram_ex \%>\% plot()
histogram_ex \%>\% geometry_pointificate() \%>\%	plot()

nz \%>\% plot()
nz \%>\% geometry_pointificate(close = TRUE) \%>\% plot()
}

