% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geometry_reflect}
\alias{geometry_reflect}
\title{Reflect positions around an axis.}
\usage{
geometry_reflect(geom, dir = c("x", "y"))
}
\arguments{
\item{geom}{A geometry data frame.}

\item{dir}{Direction in which to reflect. One of "x" and "y". (Specifying "x"
will flip the x values about the y axis.}
}
\description{
Reflect positions around an axis.
}
\examples{
scatter_ex \%>\% plot()
scatter_ex \%>\% geometry_reflect() \%>\% plot()
scatter_ex \%>\% geometry_reflect("y") \%>\% plot()

histogram_ex \%>\% plot()
histogram_ex \%>\% geometry_reflect() \%>\% plot()
histogram_ex \%>\% geometry_reflect("y") \%>\% plot()
histogram_ex \%>\% geometry_reflect("y") \%>\% geometry_flip() \%>\% plot()

nz \%>\% plot()
nz \%>\% geometry_reflect() \%>\% plot()
nz \%>\% geometry_reflect("y") \%>\% plot()
}

