% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geometry_rotate}
\alias{geometry_rotate}
\title{Rotate.}
\usage{
geometry_rotate(geom)
}
\arguments{
\item{geom}{A geometry data frame.}
}
\description{
Rotates 90 degrees by switching x and y positions.
}
\examples{
scatter_ex \%>\% plot()
scatter_ex \%>\% geometry_rotate() \%>\% plot()

histogram_ex \%>\% plot()
histogram_ex \%>\% geometry_rotate() \%>\% plot()
}

