% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geometry_scale}
\alias{geometry_scale}
\title{Scale stacked values.}
\usage{
geometry_scale(geom, dir = c("y", "x"), max = 1)
}
\arguments{
\item{geom}{A geometry data frame.}

\item{dir}{Direction in which to stack. "x" or "y" for rects,
only "y" for smooths, "r" or "theta" for arcs.}

\item{max}{Maximum value to scale to. Defaults to 1, except when scaling
"theta", where it defaults to 2 * pi.}
}
\description{
Scale stacked values.
}
\examples{
bar_ex \%>\% geometry_stack() \%>\% plot()
bar_ex \%>\% geometry_stack() \%>\% geometry_scale() \%>\% plot()

pies <- render_arc(mtcars, ~vs, ~am, 0, 0.1, 0, ~mpg / max(mpg) * 2 / pi)
pies \%>\% geometry_stack() \%>\% plot()
pies \%>\% geometry_stack() \%>\% geometry_scale() \%>\% plot()

disks <- render_arc(mtcars, ~vs, ~am, 0, 0.05, 0, 2 * pi)
disks \%>\% geometry_stack("r") \%>\% geometry_scale("r", 0.45) \%>\% plot()
}

