% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geometry_simplify}
\alias{geometry_simplify}
\title{Simplify a path geometry with Douglas-Peucker}
\usage{
geometry_simplify(geom, tol_prop = 0.1, tol_dist = NULL)
}
\arguments{
\item{geom}{A geometry data frame.}

\item{tol_prop,tol_dist}{Either specify the proportion of points to keep,
or a distance threshold.}
}
\description{
Simplify a path geometry with Douglas-Peucker
}
\seealso{
\code{\link{compute_dp_distance}()} to add the distance as
  an explicit column so you can experimenting with different thresholds
  more easily.
}

