% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geometry_transform}
\alias{geometry_transform}
\title{Perform a linear transformation on a geometry.}
\usage{
geometry_transform(geom, m, x = 0, y = 0)
}
\arguments{
\item{geom}{A geometry data frame.}

\item{m}{A 2x2 transformation matrix.}

\item{x,y}{The coordinates of the center to use for the transformation.}
}
\description{
Perform a linear transformation on a geometry, using a specified point as the
center.
}
\examples{
# Some transformation matrices
shear <- matrix(c(1, 0, 0.75, 1), nrow = 2)
reflect_y <- matrix(c(-1, 0, 0, 1), nrow = 2)
t <- 30 * pi / 180
rotate30 <- matrix(c(cos(t), -sin(t), sin(t), cos(t)), nrow = 2)

scatter_ex \%>\% plot()
scatter_ex \%>\% plot() \%>\% geometry_transform(shear)
scatter_ex \%>\% geometry_transform(reflect_y) \%>\% plot()
scatter_ex \%>\% geometry_transform(rotate30) \%>\% plot()

bar_ex \%>\% plot()
# Convert to geometry_polygon
bar_ex2 <- bar_ex \%>\% unique() \%>\% geometry_pointificate()
bar_ex2 \%>\% geometry_transform(shear) \%>\% plot()
bar_ex2 \%>\% geometry_transform(reflect_y) \%>\% plot()
bar_ex2 \%>\% geometry_transform(rotate30) \%>\% plot()

# Rotate about (10, 1) instead of origin
bar_ex2 \%>\% geometry_transform(rotate30, 10, 1) \%>\% plot()
}

