% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ggcomp-data}
\alias{bar_ex}
\alias{ggcomp-data}
\alias{histogram_ex}
\alias{nz}
\alias{scatter_ex}
\title{Various dataset to use for exmaples}
\format{Data frames.}
\usage{
bar_ex

scatter_ex

histogram_ex

nz
}
\description{
Various dataset to use for exmaples
}
\keyword{datasets}
\keyword{internal}

