% Generated by roxygen2 (4.0.2): do not edit by hand
\name{render_arc}
\alias{render_arc}
\title{Render an arc}
\usage{
render_arc(data, x, y, r1, r2, theta1, theta2)
}
\arguments{
\item{data}{A data frame.}

\item{x,y}{Location of arc}

\item{r1,r2}{Extent of radius}

\item{theta1,theta2}{Extent of angle (in radians).}
}
\description{
Render an arc
}
\examples{
render_arc(mtcars, ~vs, ~am, 0, 0.1, 0, ~mpg / max(mpg) * 2 * pi)
}

