% Generated by roxygen2 (4.0.2): do not edit by hand
\name{render_contour}
\alias{render_contour}
\title{Render 3d observations as contours.}
\usage{
render_contour(data, x, y, z, nbreaks = 10, breaks = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{x,y,z}{3d location of each point}

\item{nbreaks,breaks}{Either the number of breaks between the smallest and
largest z values, or the positions of the breaks.}
}
\value{
A path geometry.
}
\description{
Render 3d observations as contours.
}
\examples{
n <- 50
waves <- expand.grid(
   x = seq(-pi, pi, length = n),
   y = seq(-pi, pi, length = n)
)
r <- sqrt(waves$x ^ 2 + waves$y ^ 2)
waves$ripple <- cos(r ^ 2) * exp(-r / 6)
waves$hill <- exp(1 - r)

waves \%>\% render_contour(~x, ~y, ~hill) \%>\% plot()
waves \%>\% render_contour(~x, ~y, ~ripple) \%>\% plot()

# Show only where function crosses 0
waves \%>\% render_contour(~x, ~y, ~ripple, breaks = 0) \%>\% plot()
}

