% Generated by roxygen2 (4.0.2): do not edit by hand
\name{render_path}
\alias{render_path}
\alias{render_polygon}
\title{Render path and polygon geometries.}
\usage{
render_path(data, x, y)

render_polygon(data, x, y)
}
\arguments{
\item{data}{A data frame.}

\item{x,y}{Formulas specifying x and y positions.}
}
\description{
Render path and polygon geometries.
}
\examples{
# For paths and polygons, the x_ and y_ variables are lists of vectors
# See ?coord for more details
theta <- seq(0, 6*pi, length = 200)
r <- seq(1, 0, length = 200)
df <- data.frame(x = r * sin(theta), y = r * cos(theta))
spiral <- df \%>\% render_path(~x, ~y)

spiral
str(spiral)
spiral \%>\% plot()

nz
nz \%>\% plot()
#'
}

