% Generated by roxygen2 (4.0.2): do not edit by hand
\name{render_point}
\alias{render_path}
\alias{render_point}
\alias{render_polygon}
\title{Render a point, polygon and path geometries.}
\usage{
render_point(data, x, y)

render_polygon(data, x, y)

render_path(data, x, y)
}
\arguments{
\item{data}{A data frame.}

\item{x,y}{Formulas specifying x and y positions.}
}
\description{
These all use the same variables (\code{x_} and \code{y_}), but the
output looks rather different. Also note that each row describes a
different point, but each group in a grouped describes a different
path/polygon.
}
\examples{
render_point(mtcars, ~mpg, ~wt)
render_path(mtcars, ~mpg, ~wt)
render_polygon(mtcars, ~mpg, ~wt)
}

