% Generated by roxygen2 (4.0.2): do not edit by hand
\name{render_rect}
\alias{render_bar}
\alias{render_rect}
\alias{render_tile}
\title{Render a rect.}
\usage{
render_rect(data, x1, y1, x2, y2)

render_bar(data, x, y, width, halign = 0.5)

render_tile(data, x, y, width, height, halign = 0.5, valign = 0.5)
}
\arguments{
\item{data}{A data frame.}

\item{x1,y1,x2,y2}{Describe a rectangle by the locations of its sides.}

\item{x,y,width,height}{Describe a rectangle by location and dimension.}

\item{halign,valign}{Horizontal and vertical aligned. Defaults to 0.5,
centered.}
}
\description{
A rect is defined by the coordinates of its sides. Bars and tiles are
convenient parameterisations based on the length of the sides.
}
\examples{
# Two equivalent specifications
render_rect(mtcars, ~cyl - 0.5, ~gear - 0.5, ~cyl + 0.5, ~gear + 0.5)
render_tile(mtcars, ~cyl, ~gear, 1, 1)

mtcars \%>\%
  compute_count(~cyl) \%>\%
  render_bar(~x_, ~count_, width = 1)
.Last.value \%>\% plot()
}

