% Generated by roxygen2 (4.0.2): do not edit by hand
\name{render_ribbon}
\alias{render_area}
\alias{render_ribbon}
\title{Render a ribbon.}
\usage{
render_ribbon(data, x, y1, y2)

render_area(data, x, y2)
}
\arguments{
\item{data}{A data frame.}

\item{x,y1,y2}{x location and y interval.}
}
\description{
Render a ribbon.
}

