% Generated by roxygen2 (4.0.2): do not edit by hand
\name{render_segment}
\alias{render_segment}
\title{Render a line segment}
\usage{
render_segment(data, x1, y1, x2, y2)
}
\arguments{
\item{data}{A data frame.}

\item{x1,y1,x2,y2}{Locations of start and end points. sides.}
}
\description{
A line segment is a single straight line.
}
\examples{
df <- expand.grid(x = 1:2, y = 1:2)
a <- render_rect(df, ~x - 0.5, ~y - 0.5, ~x + 0.5, ~y + 0.5)
b <- render_segment(df, ~x - 0.5, ~y - 0.5, ~x + 0.5, ~y + 0.5)

plot(a)
plot(b, add = TRUE, col = "red", lwd = 2)
}

