% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vector_smooth}
\alias{vector_smooth}
\title{Smooth vectors.}
\usage{
vector_smooth(x, z, span = 0.1, n_bin = 1000, n_smooth = 100,
  weight = NULL)
}
\description{
Smooth vectors.
}
\examples{
x <- runif(1e4, 0, 4 * pi)
y <- sin(x) + runif(1e4, -0.5, 0.5)
plot(x, y)
smu <- vector_smooth(x, y, span = 0.25)
lines(smu$x, smu$y, type = "l", col = "red", lwd = 2)
x_grid <- seq(0, 4 * pi, length = 100)
lines(x_grid, sin(x_grid), type = "l", col = "blue", lwd = 2)
lines(x_grid, predict(loess(y ~ x), data.frame(x = x_grid)), col = "green", lwd = 2)
}

