% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bin_vector}
\alias{bin_vector}
\title{Bin vectors}
\usage{
bin_vector(x, width = 1, origin = 0, weight = NULL, closed = c("right",
  "left"), pad = FALSE, na.rm = FALSE)
}
\arguments{
\item{x}{A vector to bin}

\item{width}{The width of a bin}

\item{origin}{The left-most value for bins.}

\item{weight}{If specified, an integer vector of the same length as \code{x}
representing the number of occurances of each value in \code{x}}

\item{closed}{One of \code{"right"} or \code{"left"} indicating whether
right or left edges of bins are included in the bin.}

\item{pad}{A logical indicating whether the bins should be padded to include
an empty bin on each side.}
}
\description{
A generic and several implementations for binning vectors.
}
\keyword{internal}

