% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geometry_flip}
\alias{geometry_flip}
\title{Flip positions around an axis.}
\usage{
geometry_flip(geom, dir = c("x", "y"))
}
\arguments{
\item{geom}{A geometry data frame.}

\item{dir}{Direction in which to flip. One of "x" and "y".}
}
\description{
Flip positions around an axis.
}
\examples{
scatter <- render_point(mtcars, ~mpg, ~wt)
scatter \%>\% plot()
scatter \%>\% geometry_flip() \%>\% plot()
scatter \%>\% geometry_flip("y") \%>\% plot()

bar <- mtcars \%>\%
  compute_count(~cyl) \%>\%
  render_bar(~x_, ~count_, width = 1)
bar \%>\% plot()
bar \%>\% geometry_flip() \%>\% plot()
bar \%>\% geometry_flip("y") \%>\% plot()
bar \%>\% geometry_flip("y") \%>\% geometry_rotate() \%>\% plot()
}

