% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geometry_rotate}
\alias{geometry_rotate}
\title{Rotate a geometry about a pointclockwise, by a specified angle}
\usage{
geometry_rotate(geom, angle = NULL, x = 0, y = 0)
}
\arguments{
\item{geom}{A geometry data frame.}

\item{angle}{The angle by which to rotate, clockwise.}

\item{x,y}{The coordinates of the center of rotation.}
}
\description{
Rotate a geometry about a point clockwise, by a specified angle.
}
\examples{
scatter_ex \%>\% plot()
scatter_ex \%>\% geometry_rotate(90) \%>\% plot()
scatter_ex \%>\% geometry_rotate(75, x = 20, y = 20) \%>\% plot()

bar_ex \%>\% plot()
bar_ex \%>\% geometry_pointificate() \%>\%
  geometry_rotate(45, 6, 0.5) \%>\% plot()
}

