% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geometry_rotate}
\alias{geometry_rotate}
\title{Rotate.}
\usage{
geometry_rotate(geom)
}
\arguments{
\item{geom}{A geometry data frame.}
}
\description{
Rotates 90 degrees by switching x and y positions.
}
\examples{
scatter <- mtcars \%>\% render_point(~wt, ~mpg)
scatter \%>\% plot()
scatter \%>\% geometry_rotate() \%>\% plot()

bar <- mtcars \%>\%
  compute_count(~cyl) \%>\%
  render_bar(~x_, ~count_, width = 1)
bar \%>\% plot()
bar \%>\% geometry_rotate() \%>\% plot()
}

