% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geometry_stack}
\alias{geometry_stack}
\title{Stack objects on top of one another.}
\usage{
geometry_stack(geom, dir = c("y", "x"))
}
\arguments{
\item{geom}{A geometry data frame.}

\item{dir}{Direction in which to stack. "x" or "y" for rects,
only "y" for smooths, "r" or "theta" for arcs.}
}
\description{
Stack objects on top of one another.
}
\examples{
bar_ex \%>\% plot()
bar_ex \%>\% geometry_stack() \%>\% plot()

bar_ex \%>\% geometry_flip() \%>\% plot()
bar_ex \%>\% geometry_flip() \%>\% geometry_stack("x") \%>\% plot()

# Overlapping bars are stacked on top of the skyline:
df <- data.frame(x = 1:3, y = 1:3)
df \%>\% render_bar(~x, ~y, 2) \%>\% plot()
df \%>\% render_bar(~x, ~y, 2) \%>\% geometry_stack() \%>\% plot()

#' You can also stack arcs, in either r or theta direction
pies <- render_arc(mtcars, ~vs, ~am, 0, 0.1, 0, ~mpg / max(mpg) * 2 / pi)
pies \%>\% plot()
pies \%>\% geometry_stack() \%>\% plot()
pies \%>\% geometry_stack("r") \%>\% plot()

disks <- render_arc(mtcars, ~vs, ~am, 0, 0.05, 0, 2 * pi)
disks \%>\% geometry_stack("r") \%>\% plot()
}

