% Generated by roxygen2 (4.0.2): do not edit by hand
\name{geometry_warp}
\alias{geometry_warp}
\title{Warp a path or polygon with adaptive resampling}
\usage{
geometry_warp(geom, fun = c("polar", "identity"), tolerance = 0.001)
}
\arguments{
\item{geom}{A geometry data frame.}

\item{fun}{A warping function to use. Currently these are hard coded
because they much be implemented in C++ for performance reasons.}

\item{tolerance}{Approximation errors below this threshold will be
ignored.}
}
\description{
This recursively split each line segment into pieces until the transformed
point is less that \code{tolerance} away from the transformed line
segment.
}
\examples{
spiral <- data.frame(
  x = seq(0, 6 * pi, length = 10),
  y = seq(0, 1, length = 10)
)
path <- render_path(spiral, ~x, ~y)
path \%>\% plot() \%>\% points()

path \%>\% geometry_warp("polar") \%>\% plot()
path \%>\% geometry_warp("polar", tolerance = 0.00001) \%>\% plot() \%>\% points()
}

