% Generated by roxygen2 (4.0.2): do not edit by hand
\name{render_point}
\alias{render_point}
\alias{render_text}
\title{Render point and text geometries.}
\usage{
render_point(data, x, y)

render_text(data, x, y)
}
\arguments{
\item{data}{A data frame.}

\item{x,y}{Formulas specifying x and y positions.}
}
\description{
Render point and text geometries.
}
\examples{
render_point(mtcars, ~mpg, ~wt)
render_text(mtcars, ~mpg, ~wt)
}

