% Generated by roxygen2 (4.0.2): do not edit by hand
\name{render_point}
\alias{render_path}
\alias{render_point}
\alias{render_polygon}
\alias{render_text}
\title{Render point, polygon, path and text geometries.}
\usage{
render_point(data, x, y)

render_text(data, x, y)

render_polygon(data, x, y)

render_path(data, x, y)
}
\arguments{
\item{data}{A data frame.}

\item{x,y}{Formulas specifying x and y positions.}
}
\description{
Each row always represents one object, positioned according to the
\code{x_} and \code{y_} and variables. For paths and polygons, this
}
\details{
output looks rather different. Also note that each row describes a
different point or text, but each group in a grouped df describes a
different path/polygon.
}
\examples{
render_point(mtcars, ~mpg, ~wt)
render_text(mtcars, ~mpg, ~wt)

# Paths and polygons are more complicated. The x_ and y_ variables
# are lists of vectors
theta <- seq(0, 6*pi, length = 200)
r <- seq(1, 0, length = 200)
df <- data.frame(x = r * sin(theta), y = r * cos(theta))
spiral <- df \%>\% render_path(~x, ~y)
spiral
str(spiral)
spiral \%>\% plot()

nz
nz \%>\% plot()
}

