% Generated by roxygen2 (4.0.2): do not edit by hand
\name{render_segment}
\alias{render_segment}
\alias{render_spoke}
\title{Render a line segment}
\usage{
render_segment(data, x1, y1, x2, y2)

render_spoke(data, x, y, theta, r)
}
\arguments{
\item{data}{A data frame.}

\item{x1,y1,x2,y2}{Locations of start and end points.}

\item{x,y,r,theta}{Location of x points, radius and angle.}
}
\description{
A line segment is a single straight line. \code{render_spoke} is an
alternative parameterisation in terms of start point, angle and distance.
}
\examples{
df <- expand.grid(x = 1:2, y = 1:2)
a <- render_rect(df, ~x - 0.5, ~y - 0.5, ~x + 0.5, ~y + 0.5)
b <- render_segment(df, ~x - 0.5, ~y - 0.5, ~x + 0.5, ~y + 0.5)

plot(a)
plot(b, add = TRUE, col = "red", lwd = 2)

# Spokes are just an alternative parameterisation
df \%>\% render_spoke(~x, ~y, ~runif(4, 0, 2 * pi), ~0.25) \%>\% plot()
}

